feat: extract notification logic to BaseNotification
